refactor: Migrate core-ui to core-base-ui and cleanup dependencies
This commit introduces a new module `:core-base:ui` and migrates existing UI-related components from the `:core:ui` module to the new module.

Key changes include:
- Renaming and moving `core/ui` classes such as `ShareUtils`, `BaseViewModel`, `BackgroundEvent`, `EventsEffect`, `LifecycleEventEffect`, and `Transition` to the new `:core-base:ui` module.
- Creating a new `core-base:ui/build.gradle.kts` file and configuring it with relevant dependencies including Compose, Coil, and FileKit.
- Adding new platform-specific implementations for `ReportDrawnExt` and `JankStatsExtension` within the `:core-base:ui` module for Android, Desktop, JS, Native, and WasmJs.
- Moving the `ShareUtils` platform-specific implementations to the new `:core-base:ui` module.
- Removing the old `:testing` module and its `build.gradle.kts`.
- Updating the `cmp-android` module's `build.gradle.kts` and dependency tree to include the new `:core-base:ui` module and remove the old `:core:ui` dependency. Also added Coil dependency to `cmp-android`.
- Updating the `feature:settings` module's `build.gradle.kts` to depend on `:core-base:ui` and removing the old `:core:ui` dependency.
- Updating the `cmp-shared` module's `build.gradle.kts` to depend on `:core-base:ui` and added `coil-kt.coil3:coil-compose-core`.
- Updating `settings.gradle.kts` to include the new `:core-base:ui` module and remove the old `:testing` module.
- Adding `consumer-rules.pro` and `.gitignore` to the new `:core-base:ui` module.
- Implementing `SingletonImageLoader.Factory` in `cmp-android/src/main/kotlin/cmp/android/app/AndroidApp.kt` and setting up Coil with disk and memory caches.
- Providing the `ImageLoader` via `CompositionLocalProvider` in `cmp-shared/src/commonMain/kotlin/cmp/shared/SharedApp.kt`.
- Removing the old `ShareUtils.setActivityProvider` call from `cmp-android/src/main/kotlin/cmp/android/app/MainActivity.kt`.
- Creating extensions for ImageLoader, String capitalization, and SharedElement within `:core-base:ui`.
- Removing `compose.components.resources` dependency from `core-base/platform/build.gradle.kts`.

This refactoring consolidates UI components into a base module for better organization and reusability across different parts of the application, while also removing unused modules and dependencies.
