+++ cmp-android/dependencies/prodReleaseRuntimeClasspath.tree.txt
@@ -1565,7 +1565,6 @@
 |    |    |    |    |              +--- org.jetbrains.compose.ui:ui:1.7.3 (*)
 |    |    |    |    |              +--- org.jetbrains.compose.ui:ui-text:1.7.3 (*)
 |    |    |    |    |              \--- org.jetbrains.compose.ui:ui-util:1.7.3 (*)
-|    |    |    |    +--- org.jetbrains.compose.components:components-resources:1.7.0-rc01 (*)
 |    |    |    |    \--- org.jetbrains.kotlin:kotlin-parcelize-runtime:2.1.0 (*)
 |    |    |    +--- com.mohamedrejeb.calf:calf-permissions:0.7.1 (*)
 |    |    |    +--- org.jetbrains.compose.ui:ui:1.7.0-rc01 -> 1.7.3 (*)
@@ -1626,14 +1625,33 @@
 |    |    |    +--- org.jetbrains.androidx.core:core-bundle:1.0.1 (*)
 |    |    |    +--- org.jetbrains.androidx.navigation:navigation-compose:2.8.0-alpha10 (*)
 |    |    |    +--- org.jetbrains.kotlinx:kotlinx-collections-immutable:0.3.8 (*)
+|    |    |    +--- project :core:datastore (*)
+|    |    |    +--- project :core:model (*)
+|    |    |    +--- project :core-base:ui
+|    |    |    |    +--- androidx.metrics:metrics-performance:1.0.0-beta02 (*)
+|    |    |    |    +--- androidx.browser:browser:1.8.0 (*)
+|    |    |    |    +--- androidx.compose.runtime:runtime -> 1.7.6 (*)
+|    |    |    |    +--- org.jetbrains.kotlin:kotlin-stdlib:2.1.0 -> 2.1.10 (*)
+|    |    |    |    +--- io.insert-koin:koin-bom:4.0.1-RC1 (*)
+|    |    |    |    +--- io.insert-koin:koin-core:4.0.1-RC1 (*)
+|    |    |    |    +--- io.insert-koin:koin-annotations:1.4.0-RC4 (*)
+|    |    |    |    +--- org.jetbrains.androidx.lifecycle:lifecycle-viewmodel-compose:2.8.3 (*)
+|    |    |    |    +--- org.jetbrains.androidx.lifecycle:lifecycle-viewmodel:2.8.3 -> 2.8.4 (*)
+|    |    |    |    +--- org.jetbrains.androidx.lifecycle:lifecycle-viewmodel-savedstate:2.8.3 (*)
+|    |    |    |    +--- org.jetbrains.androidx.navigation:navigation-compose:2.8.0-alpha10 (*)
+|    |    |    |    +--- io.coil-kt.coil3:coil:3.0.4 (*)
+|    |    |    |    +--- io.coil-kt.coil3:coil-compose-core:3.0.4 (*)
+|    |    |    |    +--- org.jetbrains.compose.material3:material3:1.7.0-rc01 (*)
+|    |    |    |    +--- org.jetbrains.compose.components:components-resources:1.7.0-rc01 (*)
+|    |    |    |    +--- org.jetbrains.compose.components:components-ui-tooling-preview:1.7.0-rc01 (*)
+|    |    |    |    +--- io.github.vinceglb:filekit-compose:0.8.7 (*)
+|    |    |    |    \--- io.github.vinceglb:filekit-core:0.8.7 (*)
 |    |    |    +--- org.jetbrains.compose.ui:ui:1.7.0-rc01 -> 1.7.3 (*)
 |    |    |    +--- org.jetbrains.compose.foundation:foundation:1.7.0-rc01 -> 1.7.3 (*)
 |    |    |    +--- org.jetbrains.compose.material3:material3:1.7.0-rc01 (*)
 |    |    |    +--- org.jetbrains.compose.components:components-resources:1.7.0-rc01 (*)
 |    |    |    +--- org.jetbrains.compose.components:components-ui-tooling-preview:1.7.0-rc01 (*)
 |    |    |    +--- org.jetbrains.kotlinx:kotlinx-serialization-json:1.7.3 (*)
-|    |    |    +--- project :core:datastore (*)
-|    |    |    +--- project :core:model (*)
 |    |    |    \--- org.jetbrains.kotlin:kotlin-parcelize-runtime:2.1.0 (*)
 |    |    +--- org.jetbrains.compose.material3:material3:1.7.0-rc01 (*)
 |    |    +--- org.jetbrains.compose.foundation:foundation:1.7.0-rc01 -> 1.7.3 (*)
@@ -1642,9 +1660,12 @@
 |    |    +--- org.jetbrains.compose.components:components-resources:1.7.0-rc01 (*)
 |    |    \--- dev.chrisbanes.material3:material3-window-size-class-multiplatform:0.5.0 (*)
 |    +--- org.jetbrains.compose.components:components-resources:1.7.0-rc01 (*)
-|    \--- project :core-base:platform (*)
+|    +--- project :core-base:platform (*)
+|    +--- project :core-base:ui (*)
+|    \--- io.coil-kt.coil3:coil-compose-core:3.0.4 (*)
 +--- project :core:ui (*)
 +--- project :core-base:platform (*)
++--- project :core-base:ui (*)
 +--- androidx.core:core-ktx:1.16.0 (*)
 +--- androidx.appcompat:appcompat:1.7.0 (*)
 +--- androidx.activity:activity-compose:1.10.1 (*)
@@ -1660,4 +1681,5 @@
 +--- io.insert-koin:koin-compose-viewmodel:4.0.1-RC1 (*)
 +--- com.google.android.play:app-update-ktx:2.1.0 (*)
 +--- com.google.android.play:app-update:2.1.0 (*)
++--- io.coil-kt.coil3:coil:3.0.4 (*)
 \--- androidx.compose.runtime:runtime -> 1.7.6 (*)
