+++ cmp-android/src/main/kotlin/cmp/android/app/AndroidApp.kt
@@ -11,6 +11,14 @@ package cmp.android.app
 
 import android.app.Application
 import cmp.shared.utils.initKoin
+import coil3.ImageLoader
+import coil3.PlatformContext
+import coil3.SingletonImageLoader
+import coil3.disk.DiskCache
+import coil3.disk.directory
+import coil3.memory.MemoryCache
+import coil3.request.CachePolicy
+import coil3.request.crossfade
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 
@@ -22,7 +30,7 @@ import org.koin.android.ext.koin.androidLogger
  * @constructor Create empty Android app
  * @see Application
  */
-class AndroidApp : Application() {
+class AndroidApp : Application(), SingletonImageLoader.Factory {
     override fun onCreate() {
         super.onCreate()
         initKoin {
@@ -30,4 +38,23 @@ class AndroidApp : Application() {
             androidLogger() // Enables Koin's logging for debugging
         }
     }
+
+    override fun newImageLoader(context: PlatformContext): ImageLoader = ImageLoader
+        .Builder(context)
+        .networkCachePolicy(CachePolicy.ENABLED)
+        .diskCachePolicy(CachePolicy.ENABLED)
+        .memoryCachePolicy(CachePolicy.ENABLED)
+        .diskCache {
+            DiskCache.Builder()
+                .directory(context.cacheDir.resolve("image_cache"))
+                .maxSizePercent(0.25)
+                .build()
+        }
+        .memoryCache {
+            MemoryCache.Builder()
+                .maxSizePercent(context, 0.25)
+                .build()
+        }
+        .crossfade(true)
+        .build()
 }
