+++ core-base/datastore/build.gradle.kts
@@ -29,7 +29,6 @@ kotlin {
 
         commonTest.dependencies {
             implementation(libs.multiplatform.settings.test)
-            implementation(projects.testing)
         }
     }
 }
