+++ core-base/ui/src/androidMain/kotlin/template/core/base/ui/JankStatsExtensions.kt
@@ -1,0 +1,86 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
+ */
+package template.core.base.ui
+
+import androidx.compose.foundation.gestures.ScrollableState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.DisposableEffectResult
+import androidx.compose.runtime.DisposableEffectScope
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.snapshotFlow
+import androidx.compose.ui.platform.LocalView
+import androidx.metrics.performance.PerformanceMetricsState
+import androidx.metrics.performance.PerformanceMetricsState.Holder
+import kotlinx.coroutines.CoroutineScope
+
+/**
+ * Retrieves [PerformanceMetricsState.Holder] from current [LocalView] and
+ * remembers it until the View changes.
+ * @see PerformanceMetricsState.getHolderForHierarchy
+ */
+@Composable
+fun rememberMetricsStateHolder(): Holder {
+    val localView = LocalView.current
+
+    return remember(localView) {
+        PerformanceMetricsState.getHolderForHierarchy(localView)
+    }
+}
+
+/**
+ * Convenience function to work with [PerformanceMetricsState] state. The side effect is
+ * re-launched if any of the [keys] value is not equal to the previous composition.
+ * @see TrackDisposableJank if you need to work with DisposableEffect to cleanup added state.
+ */
+@Composable
+fun TrackJank(
+    vararg keys: Any,
+    reportMetric: suspend CoroutineScope.(state: Holder) -> Unit,
+) {
+    val metrics = rememberMetricsStateHolder()
+    LaunchedEffect(metrics, *keys) {
+        reportMetric(metrics)
+    }
+}
+
+/**
+ * Convenience function to work with [PerformanceMetricsState] state that needs to be cleaned up.
+ * The side effect is re-launched if any of the [keys] value is not equal to the previous composition.
+ */
+@Composable
+fun TrackDisposableJank(
+    vararg keys: Any,
+    reportMetric: DisposableEffectScope.(state: Holder) -> DisposableEffectResult,
+) {
+    val metrics = rememberMetricsStateHolder()
+    DisposableEffect(metrics, *keys) {
+        reportMetric(this, metrics)
+    }
+}
+
+/**
+ * Track jank while scrolling anything that's scrollable.
+ */
+@Composable
+actual fun TrackScrollJank(scrollableState: ScrollableState, stateName: String) {
+    TrackJank(scrollableState) { metricsHolder ->
+        snapshotFlow { scrollableState.isScrollInProgress }.collect { isScrollInProgress ->
+            metricsHolder.state?.apply {
+                if (isScrollInProgress) {
+                    putState(stateName, "Scrolling=true")
+                } else {
+                    removeState(stateName)
+                }
+            }
+        }
+    }
+}
