+++ core-base/ui/src/jsMain/kotlin/template/core/base/ui/JankStatsExtension.js.kt
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
+ */
+package template.core.base.ui
+
+import androidx.compose.foundation.gestures.ScrollableState
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun TrackScrollJank(
+    scrollableState: ScrollableState,
+    stateName: String,
+) {
+}
