+++ core-base/ui/src/nativeMain/kotlin/template/core/base/ui/ShareUtils.native.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
+ */
+package template.core.base.ui
+
+import androidx.compose.ui.graphics.ImageBitmap
+import platform.UIKit.UIActivityViewController
+import platform.UIKit.UIApplication
+
+actual object ShareUtils {
+    actual fun shareText(text: String) {
+        val currentViewController = UIApplication.sharedApplication().keyWindow?.rootViewController
+        val activityViewController = UIActivityViewController(listOf(text), null)
+        currentViewController?.presentViewController(
+            viewControllerToPresent = activityViewController,
+            animated = true,
+            completion = null,
+        )
+    }
+
+    actual suspend fun shareImage(title: String, image: ImageBitmap) {
+    }
+
+    actual suspend fun shareImage(title: String, byte: ByteArray) {
+    }
+}
