+++ settings.gradle.kts
@@ -56,9 +56,9 @@ include(":feature:home")
 include(":feature:profile")
 include(":feature:settings")
 include(":core:database")
-include(":testing")
 
 include(":core-base:datastore")
 include(":core-base:database")
 include(":core-base:network")
 include(":core-base:platform")
+include(":core-base:ui")
